Started Jump Mechanics
Added Jump() and IsGrounded() methods to SeanPlayerController and made a tunable jumpForce variable that we can modify once we figure out how to make the physics less floaty.

In the prefab I also added a Sphere collider for detecting the ground, which can be shrunk if it ends up causing unwanted collisions. That doesn't need to be assigned in the component since its found in Start()